Offer func
Fixed offer mapping so instead of working as intended on the first post request and only on the first post request, now we register the controller as a singleton service for the app, so we have one active instance of the offerController through-out the lifetime of the application.

Fixed a grammar error in the response message, added a todo comment to fix mockprice when we get the real market price

Adjusted tests to not test from the testable objects properties, but to test from pre-defined values.